+++ lib/teamtailor/client.rb
@@ -54,9 +54,7 @@ module Teamtailor
     end
 
     def jobs(page: 1, include: [], filters: {})
-      filter_params = filters.keys.map do |key|
-        { "filter[#{key}]" => filters[key] }
-      end
+      filter_params = filters.transform_keys { |key| "filter[#{key}]" }
 
       Teamtailor::Request.new(
         base_url: base_url,
@@ -67,7 +65,7 @@ module Teamtailor
           "page[number]" => page,
           "page[size]" => 30,
           "include" => include.join(","),
-        }.merge(*filter_params)
+        }.merge(filter_params)
       ).call
     end